requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
statsmodels==0.14.0
prophet==1.1.4
pmdarima==2.0.3
gunicorn==21.2.0
Werkzeug==2.3.7
python-dateutil==2.8.2
openpyxl==3.1.2
scikit-learn==1.3.0
firebase-admin==6.2.0
PyJWT==2.8.0
